fix(browser): address remaining extension feedback
