+++ Sources/AppDelegate.swift
@@ -14815,48 +14815,6 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
         handleConfiguredShortcutKeyEquivalent(event)
     }
 
-    /// Extension manifest commands run only after configured cmux shortcuts decline the event.
-    func shouldOfferBrowserWebExtensionCommand(_ event: NSEvent) -> Bool {
-        let shortcutContext = shortcutEventFocusContext(event).shortcutContext
-        for action in KeyboardShortcutSettings.Action.allCases {
-            guard action != .showHideAllWindows,
-                  action != .globalSearch,
-                  !action.isBrowserContentShortcut,
-                  KeyboardShortcutSettings.effectiveWhenClause(for: action).evaluate(shortcutContext) else {
-                continue
-            }
-            if action.usesNumberedDigitMatching {
-                if numberedConfiguredShortcutDigit(event: event, action: action) != nil {
-                    return false
-                }
-                continue
-            }
-            if configuredShortcutClaimsWebExtensionCommand(
-                event: event,
-                shortcut: KeyboardShortcutSettings.shortcut(for: action)
-            ) {
-                return false
-            }
-        }
-
-        let context = preferredMainWindowContextForShortcutRouting(event: event)
-        return !configuredCmuxShortcutActions(for: context).contains { action in
-            guard let shortcut = action.shortcut else { return false }
-            return configuredShortcutClaimsWebExtensionCommand(event: event, shortcut: shortcut)
-        }
-    }
-
-    private func configuredShortcutClaimsWebExtensionCommand(
-        event: NSEvent,
-        shortcut: StoredShortcut
-    ) -> Bool {
-        guard !shortcut.isUnbound else { return false }
-        if activeConfiguredShortcutChordPrefixForCurrentEvent == nil, shortcut.hasChord {
-            return matchShortcutStroke(event: event, stroke: shortcut.firstStroke)
-        }
-        return matchConfiguredShortcut(event: event, shortcut: shortcut)
-    }
-
     /// Route AppKit key-equivalent fallbacks through the same configured shortcut
     /// dispatcher as the local key monitor before any stale menu item can run.
     @discardableResult
@@ -15129,7 +15087,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
         return false
     }
 
-    private func matchConfiguredShortcut(event: NSEvent, shortcut: StoredShortcut) -> Bool {
+    func matchConfiguredShortcut(event: NSEvent, shortcut: StoredShortcut) -> Bool {
         guard !shortcut.isUnbound else { return false }
         if let prefix = activeConfiguredShortcutChordPrefixForCurrentEvent {
             guard let secondStroke = shortcut.secondStroke,
@@ -15172,7 +15130,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCent
         }
     }
 
-    private func numberedConfiguredShortcutDigit(
+    func numberedConfiguredShortcutDigit(
         event: NSEvent,
         action: KeyboardShortcutSettings.Action
     ) -> Int? {
